requirements.txt
biopython==1.84
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pysam==0.22.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
